Decrease likelihood of dependencies being uploaded after dependents
When uploading a folder of files, it is possible for dependencies to be
uploaded before files that depend on them. An example is a CSS file that
references a raw asset or a template that uses a module. This decreases
the liklihood that occurs through uploading file types that are often
dependencies first. This is not a perfect solution and something that
should be revisited through a bulk upload endpoint that can validate
after receiving all assets.